Copy existing .envrc if found

+++ 2022/init.zsh
@@ -14,8 +14,17 @@ echo "Created new Rust project in $folder ..."
 
 cp Makefile_template $folder/Makefile
 sed -ri "s|__aoc_url__|https://adventofcode.com/2022/day/$day/input|" $folder/Makefile
-cp envrc_template $folder/.envrc
-direnv allow $folder/.envrc
+
+local existing_envrc=$( fd --hidden --no-ignore .envrc | head -n1 )
+if [[ -n $existing_envrc ]]; then
+  echo "copying existing .envrc from '$existing_envrc' ..."
+  cp $existing_envrc $folder
+else
+  echo 'creating new .envrc ...'
+  cp envrc_template $folder/.envrc
+  direnv allow $folder/.envrc
+fi
+
 
 mkdir -p $folder/src/bin
 for p in {1..2}; do
